requirements.txt
pandas>=2.0
numpy>=1.24
scikit-learn>=1.3
scipy>=1.11
pyyaml>=6.0
tqdm>=4.65
